backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
pgvector==0.2.2
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# File Processing
python-docx==1.1.0
pypdf==3.17.4
python-magic==0.4.27

# AI/ML
openai==1.10.0
aiohttp==3.9.3

# Storage
minio==7.2.0
supabase>=2.4.0

# Redis
redis==5.0.1

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0
tenacity==8.2.3

# Validation
email-validator==2.1.0

# Logging
loguru==0.7.2
